Harden finances precision and limits

+++ app/domain/accounts/models.py
@@ -1,5 +1,5 @@
 from datetime import datetime
-from sqlalchemy import Column, Integer, String, DateTime, Float, ForeignKey
+from sqlalchemy import Column, Integer, String, DateTime, Numeric, ForeignKey
 from sqlalchemy.orm import relationship
 from app.core.database import Base
 
@@ -13,9 +13,9 @@ class Account(Base):
     user_id = Column(Integer, ForeignKey("users.id"), nullable=False)
     name = Column(String, nullable=False)
     account_type = Column(String, nullable=False)  # checking, savings, credit, etc.
-    balance = Column(Float, default=0.0)
+    balance = Column(Numeric(10, 2), default=0)
     currency = Column(String, default="USD")
-    credit_limit = Column(Float, nullable=True)
+    credit_limit = Column(Numeric(10, 2), nullable=True)
     statement_day = Column(Integer, nullable=True)  # 1-28 typically
     due_day = Column(Integer, nullable=True)
     created_at = Column(DateTime, default=datetime.utcnow)
